+++ bin/aecli.js
@@ -1,4 +1,7 @@
 #!/usr/bin/env node
+// # æternity CLI `root` file
+//
+// This script initialize all `cli` commands
 /*
  * ISC License (ISC)
  * Copyright (c) 2018 aeternity developers
@@ -16,36 +19,43 @@
  *  PERFORMANCE OF THIS SOFTWARE.
  */
 'use strict'
-
+// We'll use `commander` for parsing options
+// Also we need `esm` package to handle `ES imports`
 const program = require('commander')
 
 require = require('esm')(module/*, options */) // use to handle es6 import/export
 const utils = require('./utils/index')
 
+// Array of child command's
 const EXECUTABLE_CMD = [
   { name: 'chain', desc: 'Interact with the blockchain' },
   { name: 'inspect', desc: 'Get information on transactions, blocks,...' },
   { name: 'account', desc: 'Handle wallet operations' },
   { name: 'contract', desc: 'Compile contracts' },
   { name: 'name', desc: 'AENS system' },
-  // TODO implement oracle module
-  // {name: 'oracle', desc: 'Interact with oracles'},
+  /* TODO implement oracle module
+  * {name: 'oracle', desc: 'Interact with oracles'},
+  */
   { name: 'crypto', desc: 'Crypto helpers' }
 ]
-
+// ##Get version from `package.json`
+//
+// You get get CLI version by exec `aecli version`
 program
   .version(require('../package.json').version)
 
+// ## Initialize `config` command
 program
   .command('config')
   .description('Print the client configuration')
   .action((cmd) => utils.print.printConfig(cmd))
 
-// INIT EXEC COMMANDS
+// ## Initialize `child` command's
 utils.cli.initExecCommands(program)(EXECUTABLE_CMD)
 
-// HANDLE UNKNOWN COMMAND
+// Handle unknown command's
 program.on('command:*', () => utils.errors.unknownCommandHandler(program)(EXECUTABLE_CMD))
 
+// Parse arguments or show `help` if argument's is empty
 program.parse(process.argv)
 if (program.args.length === 0) program.help()