+++ bin/commands/inspect.js
@@ -1,4 +1,7 @@
 #!/usr/bin/env node
+// # æternity CLI `inspect` file
+//
+// This script initialize all `inspect` function
 /*
  * ISC License (ISC)
  * Copyright (c) 2018 aeternity developers
@@ -16,15 +19,6 @@
  *  PERFORMANCE OF THIS SOFTWARE.
  */
 
-//    _____                           _
-//   |_   _|                         | |
-//     | |  _ __  ___ _ __   ___  ___| |_
-//     | | | '_ \/ __| '_ \ / _ \/ __| __|
-//    _| |_| | | \__ \ |_) |  __/ (__| |_
-//   |_____|_| |_|___/ .__/ \___|\___|\__|
-//                   | |
-//                   |_|
-
 import * as R from 'ramda'
 import path from 'path'
 
@@ -38,40 +32,48 @@ import {
   printContractDescr,
   printError,
   printName,
-  printTransaction
+  printTransaction,
+  printUnderscored
 } from '../utils/print'
 import { checkPref, getBlock, readJSONFile } from '../utils/helpers'
 
+// ## Inspect function
+// That function get the param(`hash`, `height` or `name`) and show you info about it
 async function inspect (hash, option) {
   if (!hash) throw new Error('Hash required')
 
-  if (!isNaN(parseInt(hash))) {
+  // Get `block` by `height`
+  if (!isNaN(hash)) {
     await getBlockByHeight(hash, option)
     return
   }
 
   const [pref, _] = hash.split('_')
   switch (pref) {
+    // Get `block` by `hash`
     case HASH_TYPES.block:
       await getBlockByHash(hash, option)
       break
+    // Get `micro_block` by `hash`
     case HASH_TYPES.micro_block:
       await getBlockByHash(hash, option)
       break
+    // Get `account` by `hash`
     case HASH_TYPES.account:
       await getAccountByHash(hash, option)
       break
+    // Get `transaction` by `hash`
     case HASH_TYPES.transaction:
       await getTransactionByHash(hash, option)
       break
-    // case HASH_TYPES.contract:
-    //   break
+    // Get `name`
     default:
       await getName(hash, option)
       break
   }
 }
 
+// ## Inspect helper function's
 async function getBlockByHash (hash, options) {
   const { json } = options
   try {
@@ -108,10 +110,10 @@ async function getAccountByHash (hash, options) {
     const client = await initClient(options)
     await handleApiError(
       async () => {
-        const { balance, id, nonce } = await client.api.getAccountByPubkey(hash)
-        print('Account ID________________ ' + id)
-        print('Account balance___________ ' + balance)
-        print('Account nonce_____________ ' + nonce)
+        const {balance, id, nonce} = await client.api.getAccountByPubkey(hash)
+        printUnderscored('Account ID', id)
+        printUnderscored('Account balance', balance)
+        printUnderscored('Account nonce', nonce)
         print('Account Transactions: ')
         printBlockTransactions((await client.api.getPendingAccountTransactionsByPubkey(hash)).transactions, json)
       }
@@ -140,14 +142,14 @@ async function getName (name, options) {
   try {
     if (R.last(name.split('.')) !== 'aet') throw new Error('AENS TLDs must end in .aet')
     const client = await initClient(options)
-
-    printName(Object.assign(await client.api.getNameEntryByName(name), { status: 'CLAIMED' }), json)
+    const nameStatus = await client.api.getNameEntryByName(name)
+    printName(Object.assign(nameStatus, { status: 'CLAIMED' }), json)
   } catch (e) {
     if (e.response && e.response.status === 404) {
       printName({ status: 'AVAILABLE' }, json)
-      process.exit(1)
+    } else {
+      printError(e.message)
     }
-    printError(e.message)
   }
 }