+++ bin/commands/wallet.js
@@ -1,4 +1,7 @@
 #!/usr/bin/env node
+// # æternity CLI `account` file
+//
+// This script initialize all `account` function
 /*
  * ISC License (ISC)
  * Copyright (c) 2018 aeternity developers
@@ -16,34 +19,26 @@
  *  PERFORMANCE OF THIS SOFTWARE.
  */
 
-//   __          __   _ _      _
-//   \ \        / /  | | |    | |
-//    \ \  /\  / /_ _| | | ___| |_ ___
-//     \ \/  \/ / _` | | |/ _ \ __/ __|
-//      \  /\  / (_| | | |  __/ |_\__ \
-//       \/  \/ \__,_|_|_|\___|\__|___/
-//
-//
-
-import * as R from 'ramda'
-
-import { generateSecureWallet, generateSecureWalletFromPrivKey, getWalletByPathAndDecrypt } from '../utils/account'
+import { generateSecureWallet, generateSecureWalletFromPrivKey } from '../utils/account'
 import { HASH_TYPES } from '../utils/constant'
-import { initClient } from '../utils/cli'
+import { initClientByWalletFile } from '../utils/cli'
 import { handleApiError } from '../utils/errors'
 import { print, printError, printTransaction } from '../utils/print'
 import { checkPref } from '../utils/helpers'
 
+// ## Spend function
+// this function allow you to `send` token's to another `account`
 async function spend (walletPath, receiver, amount, options) {
-  let { ttl, password, json } = options
+  let { ttl, json, nonce } = options
   ttl = parseInt(ttl)
+  nonce = parseInt(nonce)
   try {
     checkPref(receiver, HASH_TYPES.account)
-    const keypair = await getWalletByPathAndDecrypt(walletPath, { password })
-    const client = await initClient(R.merge(options, { keypair }))
+    // Get `keyPair` by `walletPath`, decrypt using password and initialize `Ae` client with this `keyPair`
+    const client = await initClientByWalletFile(walletPath, options)
 
     await handleApiError(async () => {
-      let tx = await client.spend(parseInt(amount), receiver, { ttl })
+      let tx = await client.spend(parseInt(amount), receiver, { ttl, nonce })
       // if waitMined false
       if (typeof tx !== 'object') {
         tx = await client.tx(tx)
@@ -57,11 +52,12 @@ async function spend (walletPath, receiver, amount, options) {
   }
 }
 
+// ## Get `balace` function
+// This function allow you retrieve account `balance`
 async function getBalance (walletPath, options) {
-  const { password, json } = options
   try {
-    const keypair = await getWalletByPathAndDecrypt(walletPath, { password })
-    const client = await initClient(R.merge(options, { keypair }))
+    // Get `keyPair` by `walletPath`, decrypt using password and initialize `Ae` client with this `keyPair`
+    const client = await initClientByWalletFile(walletPath, options)
     await handleApiError(
       async () => print('Your balance is: ' + (await client.balance(await client.address())))
     )
@@ -70,20 +66,28 @@ async function getBalance (walletPath, options) {
   }
 }
 
+// ## Get `address` function
+// This function allow you retrieve account `public` and `private` keys
 async function getAddress (walletPath, options) {
-  const { password, privateKey } = options
+  const { privateKey } = options
   try {
-    const keypair = await getWalletByPathAndDecrypt(walletPath, { password, privateKey })
-    const client = await initClient(R.merge(options, { keypair }))
+    // Get `keyPair` by `walletPath`, decrypt using password and initialize `Ae` client with this `keyPair`
+    const { client, keypair } = await initClientByWalletFile(walletPath, options, true)
 
     await handleApiError(
-      async () => print('Your address is: ' + await client.address())
+      async () => {
+        print('Your address is: ' + await client.address())
+        if (privateKey)
+          print('Your private key is: ' + keypair.priv)
+      }
     )
   } catch (e) {
     printError(e.message)
   }
 }
 
+// ## Create secure `wallet` file
+// This function allow you to generate `keypair` and write it to secure `ethereum` like key-file
 async function createSecureWallet (walletPath, { output, password }) {
   try {
     await generateSecureWallet(walletPath, { output, password })
@@ -92,6 +96,8 @@ async function createSecureWallet (walletPath, { output, password }) {
   }
 }
 
+// ## Create secure `wallet` file from `private-key`
+// This function allow you to generate `keypair` from `private-key` and write it to secure `ethereum` like key-file
 async function createSecureWalletByPrivKey (walletPath, priv, { output, password }) {
   try {
     await generateSecureWalletFromPrivKey(walletPath, priv, { output, password })