+++ es/ae/aepp.js
@@ -26,7 +26,10 @@ import Ae from './'
 import Aens from './aens'
 import Contract from './contract'
 import Rpc from '../rpc/client'
+import Transaction from '../tx/tx'
+import Chain from '../chain'
+import Account from '../account/memory'
 
-const Aepp = Ae.compose(Contract, Aens, Rpc)
+const Aepp = Ae.compose(Account, Contract, Chain, Transaction, Aens, Rpc)
 
 export default Aepp