+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 All notable changes to this project will be documented in this file. This change
 log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [0.24.0-0.1.0]
+### Added
+- Full support of [Epoch-0.24.0](https://github.com/aeternity/epoch/releases/tag/v0.24.0)
+- (CLI) Develop `decode base58` address command in `crypto` module
+- (CLI) Add `nonce` param to all tx command's
+- (CLI) Add `gas` param to `deploy` and `call` commands
+- Add ability to create `spend` transaction natively
+- Implement `ethereum keystore` using `AES-126-CTR` and `SCRYPT` as key derivation function
+
+### Changed
+- (CLI) Change `--privateKey` to `flag` on `ACCOUNT ADDRESS` command
+- Change `node version` in `Dockerfile`
+- API endpoints to meet new Epoch specifications
+- Update `docco` config and change `rename` package to `recursive-rename`
+- Improved documentation
+
+### Removed
+- Support for < 0.24.0
+
+### Notes and known Issues
+- `ethereum keystore` usage will be removed in the next release
+- CLI will move to a separate package
+
+
 ## [0.22.0-0.1.0-beta.1]
 ### Added
 - Add **CLI** implementation
@@ -9,6 +33,8 @@ log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Changed
 - API endpoints to meet new Epoch specifications
+- Add Nonce calculation on SDK side
+- Add check for MAX_GAS in call and deploy contract
 - change hash prefix separator from $ to _
 - Add keywords ('SDK', 'CLI') to package.json
 - Link aecli to `./bin/aecli.js` in package.json (After "npm link" you can use CLI globally)
@@ -112,3 +138,4 @@ log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 [0.18.0-0.1.0]: https://github.com/aeternity/aepp-sdk-js/compare/v0.15.0-0.1.0...v0.18.0-0.1.0
 [0.18.0-0.1.1]: https://github.com/aeternity/aepp-sdk-js/compare/v0.18.0-0.1.0...v0.18.0-0.1.1
 [0.22.0-0.1.0-beta.1]: https://github.com/aeternity/aepp-sdk-js/compare/v0.18.0-0.1.1...v0.22.0-0.1.0-beta.1
+[0.24.0-0.1.0]: https://github.com/aeternity/aepp-sdk-js/compare/v0.22.0-0.1.0-beta.1...v0.24.0-0.1.0